// types.ts
export enum CropStage {
  CLONING = 'Clonación',
  PRE_VEGETATION = 'Pre-vegetación',
  VEGETATION = 'Vegetación',
  FLOWERING = 'Floración',
  DRYING_CURING = 'Secado y Curado',
  HARVESTED = 'Cosechado',
}
// ...
export interface HarvestData {
  geneticHarvests: {
    geneticsId: string;
    wetWeight: number;
    dryWeight: number;
    trimWeight: number;
  }[];
  totalWetWeight: number;
  totalDryWeight: number;
  totalTrimWeight: number;
  curingLog: { date: string; notes: string }[];
  lastBurpDate?: string;
}